backend: extract readerat from restic package
